Move generateFirebaseUpdateObject to utility file
